pyproject.toml
[tool.poetry]
name = "fashion-scoring-agent"
version = "0.1.0"
description = "智能穿搭分析Agent - 基于Qwen-VL-Chat API和PostgreSQL+pgvector的穿搭评分系统"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pillow = "^11.0.0"
opencv-python = "^4.8.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
redis = "^5.0.1"
aioredis = "^2.0.1"
pydantic = "^2.5.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
numpy = "^2.1.0"
scikit-learn = "^1.3.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
aiohttp = "^3.9.0"
httpx = ">=0.28.1,<1"
pgvector = "^0.2.4"
pyautogen = "^0.10.0"
autogen = "^0.9.9"
pydantic-settings = "^2.10.1"
ag2 = {extras = ["openai"], version = "^0.9.9"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
pre-commit = "^3.6.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true